airflow/run-notebooks/requirements.txt
boto3==1.23.6
cloud-sql-python-connector[pg8000]==0.9.0
cryptography==36.0.2
datastory==0.1.12
dataverk-vault==0.1.10
db-dtypes==1.0.5
dbt-oracle==1.3.0
geopandas==0.10.2
google-api-core==2.10.2
google-auth==2.6.6
google-cloud-bigquery==3.4.0
google-cloud-storage==2.6.0
google-api-python-client==2.48.0
great-expectations==0.15.28
influxdb==5.3.1
jira-commonmark==0.1.8
knada-gsm==0.1.2
laundromat==1.1.5
nltk==3.7
openpyxl==3.0.10
papermill==2.3.4
plotly-express==0.4.1
plotly==5.8.0
pyarrow==8.0.0
pyjwt==2.6.0
pyodbc==4.0.32
pysftp==0.2.9
python-google-places==1.4.2
scikit-learn==1.1.3
sqlalchemy==1.4.36
vega==3.6.0
wordcloud==1.8.2.2
xlrd==2.0.1
xmlschema==1.11.1
icecream==2.1.3
